+++ guillotina_dbusers/interfaces.py
@@ -3,4 +3,4 @@ from zope.interface import Interface
 
 
 class IDBUsersLayer(Interface):
-    """Marker interface layer for zodb users"""
+    """Marker interface layer for db users"""